Add files via upload
This is a script with several ML algorithms to analyze and predict breast cancer occurrence in a given population. Data used in this project is held under open-access and comes from Institute of Oncology, University Medical Centre, Ljubljana, Yugoslavia - Zwitter, M. & Soklic, M. (1988). https://doi.org/10.24432/C51P4M. Three models were applied: Random Forest, Pruned Forest and Neural Network.  Small visualizations and reports included in the code. This was done as a part of small semestral-assignment for my Postgraduate Degree in Artificial Intelligence in Data Analysis (University of Economics in Katowice). I upload it, because it received a good grade.

Dislaimers:

- You might want to edit file paths to suit your own personal computing machine
- You might want to translate my step-by-step comments from Polish to English.